requirements.txt
# py -m pip install -r requirements.txt --upgrade

pip
setuptools
wheel
cffi>=1.15.1
click>=8.1.3
colorama>=0.4.5
commonmark>=0.9.1
filetype>=1.0.13
imagequant>=1.0.3
Pillow>=9.2.0
pycparser>=2.21
Pygments>=2.12.0
rich>=12.4.4
typer>=0.4.2
# numpy>=1.22.4
# libtiff>=0.4.2

